+++ src/extension/prefdialog/parameter-float.cpp
@@ -84,7 +84,7 @@ ParamFloat::ParamFloat(Inkscape::XML::Node *xml, Inkscape::Extension::Extension 
  *
  * @param  in   The value to set to.
  */
-float ParamFloat::set(float in)
+double ParamFloat::set(double in)
 {
     _value = in;
     if (_value > _max) {
@@ -103,9 +103,7 @@ float ParamFloat::set(float in)
 std::string ParamFloat::value_to_string() const
 {
     char value_string[G_ASCII_DTOSTR_BUF_SIZE];
-    // TODO: Some strange rounding is going on here, resulting in parameter values quite different
-    //       from the original string value. Needs some investigation to make it less bad.
-    //       See also https://gitlab.gnome.org/GNOME/glib/issues/964
+    // TODO: Round off to avoid numeric errors in output string?
     g_ascii_dtostr(value_string, G_ASCII_DTOSTR_BUF_SIZE, _value);
     return value_string;
 }
